+++ src/query/range_query.rs
@@ -274,6 +274,7 @@ impl RangeWeight {
 }
 
 impl Weight for RangeWeight {
+
     fn scorer(&self, reader: &SegmentReader) -> Result<Box<Scorer>> {
         let max_doc = reader.max_doc();
         let mut doc_bitset = BitSet::with_max_value(max_doc);