+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Tantivy 0.7
   greatly improving performance
 - Tantivy error now rely on the failure crate (@drusellers)
 - Added support for `AND`, `OR`, `NOT` syntax in addition to the `+`,`-` syntax
-
+- Added a snippet generator with highlight (@vigneshsarma, @fulmicoton)
 
 Tantivy 0.6.1
 =========================